+++ src/lava/magma/runtime/message_infrastructure/CMakeLists.txt
@@ -30,6 +30,7 @@ set(MESSAGE_INFRASTRUCTURE_SRCS
   "message_infrastructure/csrc/core/multiprocessing.cc"
   "message_infrastructure/csrc/core/abstract_port_implementation.cc"
   "message_infrastructure/csrc/core/ports.cc"
+  "message_infrastructure/csrc/core/channel_factory.cc"
   "message_infrastructure/csrc/core/message_infrastructure_logging.cc"
   "message_infrastructure/csrc/actor/posix_actor.cc"
   "message_infrastructure/csrc/channel/shmem/shm.cc"
@@ -80,7 +81,7 @@ endif()
 set(MSG_LOG_LEVEL err CACHE STRING "Default vaule: err, error log only.")
 set(MSG_LOG_FILE_ENABLE 0 CACHE STRING "Default value: 0, print onto console only.")
 
-target_compile_definitions(message_infrastructure PRIVATE GRPC_CHANNEL)
+# target_compile_definitions(message_infrastructure PRIVATE GRPC_CHANNEL)
 target_compile_definitions(message_infrastructure PUBLIC
                             $<$<STREQUAL:${MSG_LOG_LEVEL},all>:MSG_LOG_LEVEL_ALL>
                             $<$<STREQUAL:${MSG_LOG_LEVEL},warn>:MSG_LOG_LEVEL_WARN>
