+++ src/lava/magma/runtime/message_infrastructure/message_infrastructure/csrc/core/channel_factory.h
@@ -26,17 +26,7 @@ class ChannelFactory {
                                 const size_t &size,
                                 const size_t &nbytes,
                                 const std::string &src_name,
-                                const std::string &dst_name) {
-    switch (channel_type) {
-      case DDSCHANNEL:
-        break;
-      case SOCKETCHANNEL:
-        return GetSocketChannel(nbytes, src_name, dst_name);
-      default:
-        return GetShmemChannel(size, nbytes, src_name, dst_name);
-    }
-    return NULL;
-  }
+                                const std::string &dst_name);
 
   friend ChannelFactory& GetChannelFactory();
 
@@ -45,15 +35,11 @@ class ChannelFactory {
                                    const int &port,
                                    const std::string &src_name,
                                    const std::string &dst_name,
-                                   const size_t &size) {
-    return std::make_shared<GrpcChannel>(url, port, src_name, dst_name, size);
-  }
+                                   const size_t &size);
 
   AbstractChannelPtr GetDefRPCChannel(const std::string &src_name,
                                       const std::string &dst_name,
-                                      const size_t &size) {
-    return std::make_shared<GrpcChannel>(src_name, dst_name, size);
-  }
+                                      const size_t &size);
 #endif
 
  private:
@@ -62,12 +48,7 @@ class ChannelFactory {
   static ChannelFactory channel_factory_;
 };
 
-ChannelFactory ChannelFactory::channel_factory_;
-
-ChannelFactory& GetChannelFactory() {
-  ChannelFactory &channel_factory = ChannelFactory::channel_factory_;
-  return channel_factory;
-}
+ChannelFactory& GetChannelFactory();
 
 }  // namespace message_infrastructure
